+++ CHANGELOG.rst
@@ -40,8 +40,10 @@ Normalize changelog and errata
      complex['foo'] = ["1"]  # ok
      complex['foo'].append("bar")  # raises a CoercionError
 
-  Be warned if using ``str`` as a type that just about anything will
-  happily coerce to a string, but that might not be what you want.
+  Be warned if using ``str`` as a type constraint that just about
+  anything will happily coerce to a string, but that might not be what
+  you want.  Consider using ``basestring`` instead, which will never
+  coerce successfully.
 
 0.8.0 6th March 2015
 --------------------